Split diagnostics.rs (1,651 lines) into focused sub-modules (#524)
Replace monolithic diagnostics.rs with a diagnostics/ directory module:
- rejection.rs: synapse rejection tracking and reporting (458 lines)
- neuron_tracking.rs: neuron rejection tracking and reporting (417 lines)
- target_data.rs: target data structures for sample building (152 lines)
- focus_filter.rs: focus target filtering and validation (151 lines)
- mod.rs: public API, re-exports, impact scoring adapter (532 lines)

All sub-modules under 600 lines. All existing tests pass unchanged.
